+++ packages/cli/test/integration/lib/project-generator.js
@@ -249,14 +249,12 @@ module.exports = function(projGenerator, props, projectType) {
         assert.jsonFileContent('package.json', props);
         assert.fileContent([
           ['package.json', '@loopback/build'],
+          ['package.json', '"typescript"'],
+          ['package.json', '"tslint"'],
           ['tslint.json', '@loopback/tslint-config'],
           ['tsconfig.json', '@loopback/build'],
         ]);
-        assert.noFileContent([
-          ['package.json', '"typescript"'],
-          ['tslint.json', '"rules"'],
-          ['tsconfig.json', '"compilerOptions"'],
-        ]);
+        assert.noFileContent([['tslint.json', '"rules"']]);
 
         if (projectType === 'application') {
           assert.fileContent(
@@ -423,6 +421,7 @@ module.exports = function(projGenerator, props, projectType) {
         assert.noFile(['tslint.json', 'tslint.build.json']);
         assert.noFileContent([
           ['package.json', '@loopback/build'],
+          ['package.json', '"tslint"'],
           ['tsconfig.json', '@loopback/build'],
         ]);
         assert.fileContent([