+++ benchmark/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.10](https://github.com/strongloop/loopback-next/compare/@loopback/benchmark@1.1.9...@loopback/benchmark@1.1.10) (2019-02-08)
+
+
+### Bug Fixes
+
+* **benchmark:** set TypeScript's rootDir to src ([51cba45](https://github.com/strongloop/loopback-next/commit/51cba45))
+
+
+
+
+
+## [1.1.9](https://github.com/strongloop/loopback-next/compare/@loopback/benchmark@1.1.8...@loopback/benchmark@1.1.9) (2019-01-28)
+
+**Note:** Version bump only for package @loopback/benchmark
+
+
+
+
+
+## [1.1.8](https://github.com/strongloop/loopback-next/compare/@loopback/benchmark@1.1.7...@loopback/benchmark@1.1.8) (2019-01-15)
+
+**Note:** Version bump only for package @loopback/benchmark
+
+
+
+
+
+## [1.1.7](https://github.com/strongloop/loopback-next/compare/@loopback/benchmark@1.1.6...@loopback/benchmark@1.1.7) (2019-01-14)
+
+**Note:** Version bump only for package @loopback/benchmark
+
+
+
+
+
 ## [1.1.6](https://github.com/strongloop/loopback-next/compare/@loopback/benchmark@1.1.5...@loopback/benchmark@1.1.6) (2018-12-20)
 
 **Note:** Version bump only for package @loopback/benchmark