+++ examples/soap-calculator/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.3](https://github.com/strongloop/loopback-next/compare/@loopback/example-soap-calculator@1.3.2...@loopback/example-soap-calculator@1.3.3) (2019-02-08)
+
+**Note:** Version bump only for package @loopback/example-soap-calculator
+
+
+
+
+
+## [1.3.2](https://github.com/strongloop/loopback-next/compare/@loopback/example-soap-calculator@1.3.1...@loopback/example-soap-calculator@1.3.2) (2019-01-28)
+
+**Note:** Version bump only for package @loopback/example-soap-calculator
+
+
+
+
+
+## [1.3.1](https://github.com/strongloop/loopback-next/compare/@loopback/example-soap-calculator@1.3.0...@loopback/example-soap-calculator@1.3.1) (2019-01-15)
+
+**Note:** Version bump only for package @loopback/example-soap-calculator
+
+
+
+
+
+# [1.3.0](https://github.com/strongloop/loopback-next/compare/@loopback/example-soap-calculator@1.2.4...@loopback/example-soap-calculator@1.3.0) (2019-01-14)
+
+
+### Features
+
+* always include tslint and typescript in project dev-dependencies ([e0df285](https://github.com/strongloop/loopback-next/commit/e0df285))
+
+
+
+
+
 ## [1.2.4](https://github.com/strongloop/loopback-next/compare/@loopback/example-soap-calculator@1.2.3...@loopback/example-soap-calculator@1.2.4) (2018-12-20)
 
 **Note:** Version bump only for package @loopback/example-soap-calculator