+++ packages/boot/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.12](https://github.com/strongloop/loopback-next/compare/@loopback/boot@1.0.11...@loopback/boot@1.0.12) (2019-02-08)
+
+**Note:** Version bump only for package @loopback/boot
+
+
+
+
+
+## [1.0.11](https://github.com/strongloop/loopback-next/compare/@loopback/boot@1.0.10...@loopback/boot@1.0.11) (2019-01-28)
+
+**Note:** Version bump only for package @loopback/boot
+
+
+
+
+
+## [1.0.10](https://github.com/strongloop/loopback-next/compare/@loopback/boot@1.0.9...@loopback/boot@1.0.10) (2019-01-15)
+
+**Note:** Version bump only for package @loopback/boot
+
+
+
+
+
+## [1.0.9](https://github.com/strongloop/loopback-next/compare/@loopback/boot@1.0.8...@loopback/boot@1.0.9) (2019-01-14)
+
+
+### Bug Fixes
+
+* rework tslint comments disabling "no-unused-variable" rule ([a18a3d7](https://github.com/strongloop/loopback-next/commit/a18a3d7))
+
+
+
+
+
 ## [1.0.8](https://github.com/strongloop/loopback-next/compare/@loopback/boot@1.0.7...@loopback/boot@1.0.8) (2018-12-20)
 
 **Note:** Version bump only for package @loopback/boot