+++ packages/metadata/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.6](https://github.com/strongloop/loopback-next/compare/@loopback/metadata@1.0.5...@loopback/metadata@1.0.6) (2019-02-08)
+
+
+### Bug Fixes
+
+* update to the most recent lodash version ([65ee865](https://github.com/strongloop/loopback-next/commit/65ee865))
+
+
+
+
+
+## [1.0.5](https://github.com/strongloop/loopback-next/compare/@loopback/metadata@1.0.4...@loopback/metadata@1.0.5) (2019-01-28)
+
+**Note:** Version bump only for package @loopback/metadata
+
+
+
+
+
+## [1.0.4](https://github.com/strongloop/loopback-next/compare/@loopback/metadata@1.0.3...@loopback/metadata@1.0.4) (2019-01-14)
+
+
+### Bug Fixes
+
+* rework tslint comments disabling "no-unused-variable" rule ([a18a3d7](https://github.com/strongloop/loopback-next/commit/a18a3d7))
+
+
+
+
+
 ## [1.0.3](https://github.com/strongloop/loopback-next/compare/@loopback/metadata@1.0.2...@loopback/metadata@1.0.3) (2018-12-20)
 
 **Note:** Version bump only for package @loopback/metadata